requirements.txt
alembic==1.10.1
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
blis==0.7.9
Boruta==0.3
catalogue==1.0.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
colorlover==0.3.0
comm==0.1.2
contourpy==1.0.7
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
Cython==0.29.14
databricks-cli==0.17.4
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
docker==6.0.1
entrypoints==0.4
executing==1.2.0
Flask==2.2.3
fonttools==4.39.0
funcy==1.18
future==0.18.3
gensim==3.8.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.7.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kmodes==0.12.2
lightgbm==3.3.5
llvmlite==0.37.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mlflow==2.2.1
mlxtend==0.19.0
multimethod==1.9.1
murmurhash==1.0.9
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
numba==0.54.1
numexpr==2.8.4
numpy==1.20.3
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pandas-profiling==3.6.6
parso==0.8.3
patsy==0.5.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
plac==1.1.3
platformdirs==3.1.0
plotly==5.13.1
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==11.0.0
pycaret==2.3.10
pydantic==1.10.5
Pygments==2.14.0
PyJWT==2.6.0
pyLDAvis==3.3.1
pymongo==4.3.3
pynndescent==0.5.8
pyod==1.0.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305
PyYAML==5.4.1
pyzmq==25.0.0
querystring-parser==1.2.4
regex==2022.10.31
requests==2.28.2
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.12.2
shap==0.41.0
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
spacy==2.3.9
SQLAlchemy==2.0.5.post1
sqlparse==0.4.3
srsly==1.0.6
stack-data==0.6.2
statsmodels==0.13.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
textblob==0.17.1
thinc==7.4.6
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.5.0
umap-learn==0.5.3
urllib3==1.26.14
visions==0.7.5
waitress==2.1.2
wasabi==0.10.1
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wincertstore==0.2
wordcloud==1.8.2.2
ydata-profiling==4.0.0
yellowbrick==1.2.1
zipp==3.15.0
